Split editor class
